% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KaplanMeier.R
\name{KaplanMeier}
\alias{KaplanMeier}
\title{Kaplan-Meier plot}
\usage{
KaplanMeier(
  data,
  predictor,
  event,
  eventage,
  adjust_cat = NULL,
  adjust_con = NULL,
  table = TRUE,
  returns = FALSE,
  ycut = NULL,
  HR_text = NULL,
  xlabs = "Time",
  ylabs = "Survival probability",
  ystratalabs = NULL,
  ystrataname = NULL,
  timeby = 365,
  maxtime = NULL,
  main = "Kaplan-Meier Plot",
  pval = FALSE,
  rotate = TRUE,
  years = TRUE,
  hr = FALSE,
  color = FALSE,
  colors = NULL,
  reverseHR = FALSE,
  annotation = NULL,
  reverse_table = FALSE,
  No_legends = FALSE,
  No_names = FALSE,
  low_legend = FALSE
)
}
\arguments{
\item{data}{data.frame}

\item{predictor}{Predictor of interest for plotting from @data.}

\item{event}{event}

\item{eventage}{eventage}
}
\value{
A ggplot / grob
}
\description{
For creating a Kaplan-Meier plot using ggplot2
}
