% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KaplanMeier.R
\name{KaplanMeier}
\alias{KaplanMeier}
\title{Kaplan-Meier plot}
\usage{
KaplanMeier(
  data,
  predictor,
  event,
  eventage,
  table = TRUE,
  reverse_table = FALSE,
  timeby = 365,
  years = TRUE,
  hr = TRUE,
  reverseHR = FALSE,
  adjust_cat = NULL,
  adjust_con = NULL,
  returns = TRUE,
  ycut = NULL,
  xlabs = "Time",
  ylabs = "Survival probability",
  main = "Kaplan-Meier Plot",
  legend_title = NULL,
  legend_names = NULL,
  mintime = NULL,
  maxtime = NULL,
  rotate = TRUE,
  color = TRUE,
  colors = NULL,
  reverse_colors = FALSE,
  annotation = NULL,
  no_legends = FALSE,
  no_names = FALSE,
  low_legend = FALSE,
  y_tick_n = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{predictor}{Predictor of interest for plotting from @data.}

\item{event}{event}

\item{eventage}{eventage}

\item{table}{Display survival table below graph; default is TRUE.}

\item{reverse_table}{reverse the upper/lower part of table; default is FALSE.}

\item{timeby}{Cut points for the table; default is 365 (days).}

\item{years}{X-axis in years - else in days; default is TRUE.}

\item{hr}{Make and plot cox proportional HR; default is TRUE.}

\item{reverseHR}{Change case/control categories for HR; default is FALSE.}

\item{adjust_cat}{Categorical variables to adjust the cox HR for.}

\item{adjust_con}{Continues variables to adjust the cox HR for.}

\item{returns}{outputs the ggplot/grob object; default is TRUE.}

\item{ycut}{Cut y-axis at a certain percentage; default is NULL.}

\item{xlabs}{X-axis label}

\item{ylabs}{Y-axis label}

\item{main}{Plot title.}

\item{legend_title}{Legend title; default is name of predictor.}

\item{legend_names}{Define variable names for legend text.}

\item{mintime}{To get consistent min times on x axes in stratified plots.}

\item{maxtime}{To get consistent max times on x axes in stratified plots.}

\item{rotate}{Switch the predictor order; default is FALSE.}

\item{color}{Color plot, else black/greytone; default is TRUE.}

\item{colors}{Set of defined colors for predictor groups; default is NULL.}

\item{reverse_colors}{Flip the predictor group; default is FALSE.}

\item{annotation}{Paste an annotation in the corner; default is NULL.}

\item{no_legends}{Removes legend; default is FALSE.}

\item{no_names}{Removes y names; default is FALSE.}

\item{low_legend}{Put legends in the bottom, else in the top corner;}

\item{y_tick_n}{Number of y-ticks; default is NULL.}
}
\value{
A ggplot / grob
}
\description{
For creating a Kaplan-Meier plot using ggplot2
}
